chore: enterprise readiness roadmap - phases 2-10
Phase 2: CI/CD Quality Gates
- Add Pint code style job to ci-security.yml
- Update pre-commit hook to run Pint

Phase 3: PHPStan Baseline Reduction 1860 → 1120
- Add makeRequest()/logError() to BaseManufacturerService (-222 suppressions)
- Add type hints to AuthorizationService methods (-20 suppressions)
- Add @property PHPDoc to FuelMonthlyAllocation, ComponentReplacement, Machine models
- Remove invalid @method PHPDoc tags from 40 model files
- Add Builder<static> generic types to scope methods via HasTeamFilters trait
- Fix MineAreaDetail, MaintenanceAlert, BillingPortal, Reports type issues
- Fix bare return; in mixed-returning Livewire methods

Phase 4: Test Coverage Expansion
- Add AuthorizationServiceTest (22 tests)
- Add GenerateReportJobTest (6 tests)
- Add AlertGenerationJobTest (6 tests)
- Add GeofenceCrossingDetectionJobTest (6 tests)
- Add ReportFactory with processing/completed/failed states
- Add Webhook replay protection (5-minute timestamp check)

Phase 5: API Documentation
- Install dedoc/scramble for OpenAPI docs at /docs/api

Phase 6: Eager Loading Audit
- All API controllers already use eager loading (verified)

Phase 7: Observability
- Install Laravel Horizon for queue monitoring (/horizon)
- Install Laravel Pulse for application metrics
- Add HealthController at GET /health (db/cache/queue/storage checks)

Phase 8: Test Suite Expansion
- 149 tests passing (from 113), 6 skipped (zip extension)

Phase 9: GDPR and Data Retention
- Add PurgeOldAuditLogsJob, PurgeOldFeedPostsJob, PurgeExpiredSoftDeletesJob
- Schedule all retention jobs in routes/console.php with withoutOverlapping

Phase 10: Kubernetes Infrastructure
- Update k8s liveness/readiness probes from /up to /health

+++ config/sentry.php
@@ -13,5 +13,5 @@ return [
     'dsn' => env('SENTRY_DSN', null),
     'environment' => env('SENTRY_ENVIRONMENT', env('APP_ENV', 'production')),
     'release' => env('SENTRY_RELEASE', null),
-    'traces_sample_rate' => env('SENTRY_TRACES_SAMPLE_RATE', 0.0),
+    'traces_sample_rate' => (float) env('SENTRY_TRACES_SAMPLE_RATE', 0.0),
 ];
